authentication

+++ Back-end/src/main/java/fct/unl/pt/instagramplus/Controllers/Accounts/AccountsControllerClass.java
@@ -5,12 +5,15 @@ import fct.unl.pt.instagramplus.Models.Accounts.Account;
 import fct.unl.pt.instagramplus.Models.Follower;
 import fct.unl.pt.instagramplus.Models.ProfileViewer;
 import fct.unl.pt.instagramplus.Services.AccountsServices.AccountServiceClass;
+import fct.unl.pt.instagramplus.Services.Result;
 import fct.unl.pt.instagramplus.Utils.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 @RestController
 public class AccountsControllerClass implements AccountsControllerInterface{
@@ -19,56 +22,65 @@ public class AccountsControllerClass implements AccountsControllerInterface{
     AccountServiceClass accountService;
 
     @Override
-    public ResponseEntity<Long> createAccount(Account account) {
-        Logger.info("Request: CREATEACCOUNT");
-        return Response.resultOrErrorCode(accountService.createAccount(account));
-    }
-
-    @Override
-    public ResponseEntity<Account> gettAccount(Long id) {
+    public ResponseEntity<Account> getAccount(Long accountRequestId, Long id) {
         Logger.info("Request: GETACCOUNT");
+        accountService.setNewViewer(accountRequestId, id);
         return Response.resultOrErrorCode(accountService.getAccount(id));
     }
 
     @Override
-    public ResponseEntity<Void> deleteAccount(Long id) {
+    public ResponseEntity<Void> deleteAccount(Long accountRequestId,Long id) {
         Logger.info("Request: DELETEACCOUNT");
+        if(!accountRequestId.equals(id))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.deleteAccount(id));
     }
 
     @Override
-    public ResponseEntity<List<ProfileViewer>> gettAccountViweres(Long id) {
+    public ResponseEntity<List<ProfileViewer>> getAccountViweres(Long accountRequestId,Long id) {
         Logger.info("Request: ACCOUNTVIWERES");
+        if(!accountRequestId.equals(id))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.getAccountViweres(id));
     }
 
     @Override
-    public ResponseEntity<Void> followAccount(Follower follower) {
+    public ResponseEntity<Void> followAccount(Long accountRequestId,Follower follower) {
         Logger.info("Request: FOLLOWACCOUNT");
+        if(!accountRequestId.equals(follower.getAccountId()))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.followAccount(follower));
     }
 
     @Override
-    public ResponseEntity<Void> unfollowAccount(Follower follower) {
+    public ResponseEntity<Void> unfollowAccount(Long accountRequestId,Follower follower) {
         Logger.info("Request: UNFOLLOWACCOUNT");
+        if(!accountRequestId.equals(follower.getAccountId()))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.unfollowAccount(follower));
     }
 
     @Override
-    public ResponseEntity<List<Follower>> getFollowersAccount(Long id) {
+    public ResponseEntity<List<Follower>> getFollowersAccount(Long accountRequestId,Long id) {
         Logger.info("Request: ACCOUNTFOLLOWERS");
+        if(!accountRequestId.equals(id))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.getFollowersAccount(id));
     }
 
     @Override
-    public ResponseEntity<List<Follower>> getAccountFollowings(Long id) {
+    public ResponseEntity<List<Follower>> getAccountFollowings(Long accountRequestId,Long id) {
         Logger.info("Request: ACCOUNTFOLLOWINGS");
+        if(!accountRequestId.equals(id))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.getAccountFollowings(id));
     }
 
     @Override
-    public ResponseEntity<Void> changeAccountVisibility(Long id) {
+    public ResponseEntity<Void> changeAccountVisibility(Long accountRequestId,Long id) {
         Logger.info("Request: CHANGEVISIBILITY");
+        if(!accountRequestId.equals(id))
+            return ResponseEntity.status(UNAUTHORIZED).build();
         return Response.resultOrErrorCode(accountService.changeVisibility(id));
     }
